+++ Solutions/Google Apigee/ReleaseNotes.md
@@ -1,3 +1,4 @@
  **Version** | **Date Modified (DD-MM-YYYY)**| **ChangeHistory**                                                                         |
 |------------|-------------------------------|-------------------------------------------------------------------------------------------|
-| 3.0.0      | 05-09-2024                    | Updated the python runtime version to 3.11                                                |
+| 3.1.0      | 28-02-2025                    | Added new CCP **Data Connector** to the Solution                                          |
+| 3.0.0      | 05-09-2024                    | Updated the python runtime version to 3.11 in **Data Connector** Function APP             |